references should be prefered over pointers
Surveying the literature, such as it is, on the internet, I find a heavy
recommendation of using references whenever practical. The reasons
typically given include (but are not limited to):
	1) References do not require additional allocation, unlike
		pointers
	2) If a pointer is null, it's not necessarily clear what
		that means
	3) References are typically "safer" unless One goes about
		a convoluted way to "cause troubel", such as
		"int &r = *((int *)NULL);"
	4) There is simply less typing involved with accessing the
		innard of a reference to class object, making
		development faster
So, whenever practical I want to use references instead of
pointers.

I also want to make sure when declaring a reference, the '&' is
adjacent to the type to which the reference refers and not the
identifier in order to emphasize the fact the type of the
identifier is a reference to some type.