apparently needlessly complex condition
In 'UncoveredSet()', We see the use of an expression:
	!(IIsUncovered ? IIsASchwartz : true)
and also of:
	IIsUncovered ? IIsASchwartz : true

These expressions feel unnecessarily complex to Me. At the least,
it takes longer to figure out exactly what the expressions mean.

Simplifying the text of the first expression, We get:
	!(A ? B : true)
We have four possible states:
	A is true && B is true
	A is true && B is false
	A is false && B is true
	A is false && B is false

If A is false, the parenthesized expression is true and the entire
expression is false. Therefore, for the entire expression to be
true, at a minimum, A must be true, giving the beginings of a
clearer expression:
	A <some logical operator> <some expression>

Presuming A is true, if B is true, the parenthesized expression is
true and the entire expression is false and, if B is false, the
parenthesized expression is false and the entire expression is
true. Therefore, a clearer expression takes the form:
	A && !B
or, using original identifiers:
	IIsUncovered && !IIsASchwartz

Simplifying the text of the second expression, We get:
	A ? B : true
We have the same four possible states as with the first
expression.

If A is false, the entire expression is true and the beginings of
a clearer expression take the form of:
	!A || <some expression>
in order for the entire expression to be true.

Presuming A is true, if B is true, the entire expression is true
and, if B is false, the entire expression is false. Therefore a
clearer form of the expression is:
	!A || B
or, using original identifiers:
	!IIsUncovered || IIsASchwartz