harmonizing for-loop direction
In this modification, I change two for-loops so, instead of
counting down, they are counting up. While such a change does
alter the output, this change is justified, or at least
non-unjustified, because of the purpose of the loops. These two
loops attempt to find the Candidate which has not been
eliminated yet and has the lowest 'voteCountForThisRound' member
value. In the event of a tie for lowest member value, 1 of the
tied, non-eliminated, Candidates is selected. In order to have
such tie-breaking "built in" to the algorith, 'RandCandPerm[]'
is used to determine the order in which to analyze Candidates
and the 1st Candidate analyzed with the lowest member value is
selected.  Changing the loop direction from "counting down" to
"counting up" reverses the order of Candidate analysis, changing
the selected Candidate and the election results.

To illustrate, suppose Candidates were examined in an order
where member values are, respectively, 8, 6, 7, 5, 3, 0, and 9.
Reversing the order of analysis does not change the fact the
Candidate with a member of value 0 has the lowest member value
and does not change the fact the 6th analyzed Candidate is
selected by this particular algorithm. However, if the member
values were 8, 6, 7, 5, 3, 0, and 0, reversing the order does
change the selected Candidate from the 6th to the 7th.

The possibility does exist the direction of the loop counting is
significant for these modified algorithms. However, after
consulting with Warren D. Smith, no such significant factor has
presented itself. If this fact were to change, a 'repeal' of
this modification may be justified, if not required.