function heading comment changes
After much reflection, I have come to the conclusion the current
approach to providing function heading comments is generally
unworkable. The indentation and/or alignment principles by which
I have so far used often result in a large bit of unused white
space needlessly and interfere with the attempts to make the
file more readible as a result.

So, I have decided to switch to using NaturalDocs format for
function comments. Going forward, whenever I make a modification
which alters a function, I will create a separate modification
which make sure the function heading comment of that function
reflects the new style, as well as the functions appearing
immediately before and after that file.

For this particular modification, the heading changes are for
'IRV()', 'BTRIRV()', 'SimmonsCond()', 'SmithIRV()', 'Top3IRV()',
and 'Coombs()'.