harmonizing the bodies of 'BRDriver()' and 'RWBRDriver()' and adjusting their commentary
After simplifying the bodies of 'BRDriver()' and 'RWBRDriver()', We
see the two functions are very similar in structure. The only
differences in their respective structures are the number of ignorance
levels dealt with by each function and the function called in the body
of the respective for-loops.

In the creation of each election scenario, the scenarios have
different 'components' to consider. Simulated elections have the
components of:
        Voter ignorance level
        utility generation method
        voting-honestly-versus-voting-strategically level
        numbers of Voters
while "real world" elections have the components of:
        Voter ignorance level
        voting-honestly-versus-voting-strategically level
Therefore, One readily sees the two functions follow the structure of:
        set the "real world" state of the elections
        for each ignorance level to consider
                set the ignorance level for the subset of elections
                produce output in accordance with whatever the next
                   component to consider is
        enable the output file
        output the summary of the normalized regret data
        disable the output file

Therefore, We can simplify the code by calling a unified function
which takes as arguments the "real world" state of the elections, the
maximum number of ignorance levels to consider, a pointer to the
function which produces output in accordance with whatever the next
component to consider is, and a string indicating the calling
function.

I also notice I missed including parameter information in the
descriptions of the two driver functions; I will add that as well.