making 'PrintTheVotersBayesianRegret()' more readible
In trying to better understand the code both now and in the
future, 'PrintTheVotersBayesianRegret()' should be written in
such a way which more readily reflects the algorith at hand.
Generally speaking, that algorithm looks like so:
        if this election is either a real world one or has an
          acceptable number of Voters,
                set some limits and regret data
                for each number of allowable Candidates from
                  least to most,
                        do stuff
I have used a reference to help automatically update regret
data. In order to make sure that data remains unaltered when
processing real world elections, the for-loop is deliberately
set to exit after one run for real world elections. At the same
time, part of Me wonders if there is a better way to handle the
real world elections.