changing a counting-down loop to a counting-up loop
Adjusting the direction of the loop counting will make for more easily
followable code down soon. Yes, the output does differ as a result of
this change and for good reason: the value returned by the
psuedo-random number generator is applied to different Candidates and
different Voters because of the change in loop direction. However, I
see no reason why the direction should matter, leaving the ability of
the program to aid in determining which voting methods tend to result
in the lowest Bayesian regret for an entire populace generally in
tact; the only difference is the deployment of random values and not
the accuracy of algorithms.