making 'oneVoter::topDownPrefs' a standard container and resizing 'vector's of Candidates to keep them up to date
The same reasons for deploying standard C++ containers applies to this
modification as well as to prior ones doing so. This modification also
allows Me to make use of range-based for-loops and dispense with
certain parameters in the bodies of certain functions and arguments in
their respective calls, increasing readibility and comprehensibility.