removing an unnecessary reliance on 'RandCandPerm[]'
Unlike the previous change with a similar name, differences in output
are expected due to the fact the 'shuffling' of 'RandCandPerm[]'
occurs less often. As a result of the fewer shuffles, different values
are expected to be returned from the pseudo-random number generator at
different points in testing. These difference values from the
generator account for the difference in output.