consolidating Copeland scores with Candidates
Note: this mod only results in no changes in test output due to an
obscure condition in 'Maximum()' The function only tests whether a
value is equal to or greater than the current maximum if the type of
the value being examined is 'unsigned'; otherwise, the function
insists the value exceed the current maximum in order to become the
'new maximum'. The wording of the variable might lead One to think,
"Oh the type of the value is an unsigned type of some sort but not
necessarily the actual 'unsigned' type." I might change the variable
name in a future patch in order to reduce the risk of confusion.