simplifying 'TopMedianRating()' and related tweaks
The creation of a type to represent a collection of Voters,
'Constituency', helps improve readability of the code by
reducing the "visual noise" of the vector type used otherwise.

The addition of an 'index' member to the 'oneVoter' type helps
improve readability of the code by reducing the need for
external counter variables.

Creation of a 'resizeAndReset<>()' specialization for
'Constituency' objects simplifies the initialization of each
'oneVoter::index' variable.